pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgmgmt"
version = "0.1.0"
description = "Telegram management bot with bot-to-bot loop prevention."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "python-telegram-bot[rate-limiter]==21.6",
    "redis>=5.0,<6",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "structlog>=24.1",
    "uvloop>=0.19; platform_system != 'Windows'",
    "orjson>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "fakeredis[lua]>=2.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
tgmgmt = "tgmgmt.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "SIM"]
